implement graph adaptive size

+++ src/WpfUI/UI/Graph/NodeUI.cs
@@ -4,5 +4,5 @@ public class NodeUI
     public string Name { get; set; } = string.Empty;
     public double X { get; set; }
     public double Y { get; set; }
-    public int Radius { get; set; }
+    public double Radius { get; set; }
 }
